renaming frontend folders
