+++ frontend/src/components/NavStack/index.tsx
